requirements.txt
langchain==0.3.23
langchain_chroma==0.2.2
langchain_community==0.3.21
langchain_groq==0.3.2
langchain_huggingface==0.1.2
python-dotenv==1.1.0
streamlit==1.42.0
